+++ project2/utilities/curve.hpp
@@ -5,12 +5,12 @@
 
 class Curve : public Data {
 public:
-  std::vector<int> tVec;    // time vector
+  std::vector<float> tVec;    // time vector
   std::vector<float> gxVec; // grid vector (x-axis)
   std::vector<float> gyVec; // grid vector (y-axis)
   std::vector<float> key;   // LSH key
 
-  Curve(std::vector<float>, std::vector<int>, std::string);
+  Curve(std::vector<float>, std::vector<float>, std::string);
   ~Curve();
   void filter(float);
   void padding();
